Move a bunch of UI elements into their own module.
