+++ src/lib.rs
@@ -9,3 +9,4 @@ pub mod layout;
 pub mod particle;
 pub mod rect;
 pub mod screen;
+pub mod ui;
